[FIX] fixed TOPPView_test
